test(codex): split codegraph adapter coverage
