+++ script/package-registration-audit.test.ts
@@ -32,8 +32,6 @@ const mcpPackagePaths: readonly string[] = [
 ] as const
 const adapterPackagePaths: readonly string[] = ["packages/omo-codex", "packages/omo-opencode"] as const
 const skillPackagePaths: readonly string[] = ["packages/shared-skills"] as const
-const shimSourceRoots: readonly string[] = ["packages/omo-opencode/src", "packages/omo-codex/src"] as const
-const reExportShimFirstLinePattern = /^export (\*|\{).*from ["'](@oh-my-opencode\/[^/"']+)/
 
 const layerRanks = {
   skill: 1,
@@ -50,15 +48,6 @@ type PackageManifest = {
   readonly peerDependencies: Record<string, string>
 }
 
-type RootManifest = PackageManifest & {
-  readonly workspaces: readonly string[]
-}
-
-type ReExportShim = {
-  readonly path: string
-  readonly targetPackage: string
-}
-
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null && !Array.isArray(value)
 }
@@ -66,14 +55,6 @@ function isRecord(value: unknown): value is Record<string, unknown> {
 function readStringField(record: Record<string, unknown>, key: string): string {
   const value = record[key]
   if (typeof value !== "string") throw new Error(`Expected ${key} to be a string`)
-  return value
-}
-
-function readStringArrayField(record: Record<string, unknown>, key: string): readonly string[] {
-  const value = record[key]
-  if (!Array.isArray(value) || !value.every((entry) => typeof entry === "string")) {
-    throw new Error(`Expected ${key} to be a string array`)
-  }
   return value
 }
 
@@ -102,19 +83,6 @@ async function readManifest(packageJsonPath: string): Promise<PackageManifest> {
   }
 }
 
-async function readRootManifest(): Promise<RootManifest> {
-  const parsed: unknown = JSON.parse(await readFile("package.json", "utf8"))
-  if (!isRecord(parsed)) throw new Error("package.json is not a JSON object")
-  return {
-    name: readStringField(parsed, "name"),
-    workspaces: readStringArrayField(parsed, "workspaces"),
-    scripts: readStringRecordField(parsed, "scripts"),
-    dependencies: readStringRecordField(parsed, "dependencies"),
-    devDependencies: readStringRecordField(parsed, "devDependencies"),
-    peerDependencies: readStringRecordField(parsed, "peerDependencies"),
-  }
-}
-
 async function collectFiles(root: string, predicate: (path: string) => boolean): Promise<readonly string[]> {
   const entries = await readdir(root, { withFileTypes: true })
   const files: string[] = []
@@ -132,31 +100,12 @@ async function collectFiles(root: string, predicate: (path: string) => boolean):
   return files
 }
 
-async function collectTrackedFiles(roots: readonly string[], predicate: (path: string) => boolean): Promise<readonly string[]> {
-  return (await Promise.all(roots.map((root) => collectFiles(root, predicate)))).flat().toSorted()
-}
-
 async function discoverPackagePaths(): Promise<readonly string[]> {
   const packageNames = await readdir("packages")
   return packageNames
     .map((name) => `packages/${name}`)
     .filter((path) => existsSync(join(path, "package.json")))
     .toSorted()
-}
-
-async function collectReExportShims(): Promise<readonly ReExportShim[]> {
-  const files = await collectTrackedFiles(shimSourceRoots, (path) => path.endsWith(".ts"))
-
-  const shims: ReExportShim[] = []
-  for (const path of files) {
-    const source = await readFile(path, "utf8")
-    const firstLine = source.split(/\r?\n/, 1)[0] ?? ""
-    const targetPackage = reExportShimFirstLinePattern.exec(firstLine)?.[2]
-    if (targetPackage === undefined) continue
-    shims.push({ path, targetPackage })
-  }
-
-  return shims
 }
 
 function isManagedWorkspacePackage(path: string): boolean {
@@ -168,28 +117,12 @@ function isManagedWorkspacePackage(path: string): boolean {
   )
 }
 
-function isNestedCodexPluginPackage(path: string): boolean {
-  return path.startsWith("packages/omo-codex/plugin/")
-}
-
-function isRootManagedTypecheckPackage(path: string): boolean {
-  return !isNestedCodexPluginPackage(path)
-}
-
 function packageLayer(path: string): keyof typeof layerRanks | undefined {
   if (corePackagePaths.includes(path)) return "core"
   if (mcpPackagePaths.includes(path)) return "mcp"
   if (adapterPackagePaths.includes(path)) return "adapter"
   if (skillPackagePaths.includes(path)) return "skill"
   return undefined
-}
-
-function extractTypecheckPackagePaths(command: string): readonly string[] {
-  return [...command.matchAll(/tsgo --noEmit -p (packages\/[^ ]+\/tsconfig\.json)/g)]
-    .map((match) => match[1])
-    .filter((path): path is string => path !== undefined)
-    .map((path) => path.slice(0, -"/tsconfig.json".length))
-    .toSorted()
 }
 
 function extractSharedCoreGuardPackagePaths(source: string): readonly string[] {
@@ -204,36 +137,6 @@ function extractSharedCoreGuardPackagePaths(source: string): readonly string[] {
 }
 
 describe("package registration audit", () => {
-  test("#given managed packages #when root registration is audited #then workspaces typecheck and dev deps stay aligned", async () => {
-    // given
-    const root = await readRootManifest()
-    const managedPackagePaths = (await discoverPackagePaths()).filter(isManagedWorkspacePackage)
-    const managedWorkspacePaths = managedPackagePaths.filter((path) => !mcpPackagePaths.includes(path) || root.workspaces.includes(path))
-    const expectedTypecheckPaths = managedWorkspacePaths.filter((path) => existsSync(join(path, "tsconfig.json")))
-    const expectedDevDependencyNames = (
-      await Promise.all(
-        managedWorkspacePaths
-          .filter((path) => path !== "packages/omo-opencode")
-          .map((path) => readManifest(join(path, "package.json")).then((manifest) => manifest.name)),
-      )
-    ).toSorted()
-
-    // when
-    const actualWorkspacePaths = root.workspaces.filter(isManagedWorkspacePackage).toSorted()
-    const actualTypecheckPaths = extractTypecheckPackagePaths(root.scripts["typecheck:packages"] ?? "").filter(
-      isRootManagedTypecheckPackage,
-    )
-    const actualDevDependencyNames = Object.entries(root.devDependencies)
-      .filter((entry) => entry[1] === "workspace:*" && entry[0].startsWith("@oh-my-opencode/"))
-      .map((entry) => entry[0])
-      .toSorted()
-
-    // then
-    expect(actualWorkspacePaths).toEqual(managedWorkspacePaths.toSorted())
-    expect(actualTypecheckPaths).toEqual(expectedTypecheckPaths.toSorted())
-    expect(actualDevDependencyNames).toEqual(expectedDevDependencyNames)
-  })
-
   test("#given shared extraction guard #when audited #then every core package is covered", async () => {
     // given
     const guardSource = await readFile("script/shared-core-extraction-guard.test.ts", "utf8")
@@ -295,24 +198,4 @@ describe("package registration audit", () => {
     expect(reverseEdges).toEqual([])
   })
 
-  test("#given exact re-export shim scan #when inventory docs are checked #then total targets and every path are present", async () => {
-    // given
-    const docPath = "docs/reference/re-export-shim-inventory.md"
-    const [doc, shims] = await Promise.all([readFile(docPath, "utf8"), collectReExportShims()])
-
-    // when
-    const totalMatch = /Total shim exports found: (\d+)\./.exec(doc)
-    if (totalMatch?.[1] === undefined) throw new Error(`${docPath} is missing the total shim count`)
-
-    const documentedTotal = Number.parseInt(totalMatch[1], 10)
-    const targetPackages = [...new Set(shims.map((shim) => shim.targetPackage))].toSorted()
-    const missingTargetPackages = targetPackages.filter((targetPackage) => !doc.includes(`\`${targetPackage}\``))
-    const missingPaths = shims.map((shim) => shim.path).filter((path) => !doc.includes(`\`${path}\``))
-
-    // then
-    expect(doc).toContain("Re-export Shim Inventory")
-    expect(documentedTotal).toBe(shims.length)
-    expect(missingTargetPackages).toEqual([])
-    expect(missingPaths).toEqual([])
-  }, { timeout: 20_000 })
 })
